+++ src/components/sidebar/debug/debug.js
@@ -14,7 +14,19 @@
 			templateUrl: 'components/sidebar/debug/debug.html',
 			link: function ($s, $e, attrs) {
 				
+				$s.history = $history;
+				$s.version = Zemit.version;
 				
+				$s.dump = () => {
+					var changes = $history.dump();
+					var content = $zm.content.get(true, true);
+					var data = {
+						changes: changes,
+						content: content,
+						version: Zemit.version
+					};
+					console.log(data);
+				};
 			}
 		}
 	}]);