+++ src/components/sidebar/sidebar.js
@@ -34,9 +34,6 @@
 					if(config.debug && config.sidebar.tabs.debug.visible) {
 						$s.tabs.toggle('debug');
 					}
-					else if(!config.debug && !$s.tabs.showContent) {
-						$s.tabs.toggle('debug');
-					}
 					
 					config.debug = !config.debug;
 				};