+++ CHANGELOG.md
@@ -8,11 +8,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - 0.0.8:
 - Complete workspace area
 - Confirm dialog on document clearing
-- Add a widget outside a row automatically creates a row
-- Hold and drag a widget (slower delay)
 - Android back button support
 - Add next button on dimissable panels
-- Add pointer cursor on tab items
+- Update app's inner documentation
+- Bug with data cache (doesn't load on refresh)
+- 0.0.9
+- Show "prevent cursor" when dragging a widget over another that doesn't accept its type
+
+## [0.0.8]
+- Can add widgets outside of a row
+- Hold and drag a widget (slower delay)
+- Update InteractJS to 1.3.2
 
 ## [0.0.7]
 - Change opacity on mouseover on sidebar icons